Move `createMockModel` to subpath import
